% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_snails.R
\name{sim_snails}
\alias{sim_snails}
\title{Generate multiple generations of simulations of genetic drive}
\usage{
sim_snails(popsize = 50, tidepools = 3, ngen = 10)
}
\arguments{
\item{popsize}{Population size (start w half to each allele); default 50.}

\item{tidepools}{Number of populations; default 3.}

\item{ngen}{Number of generations to simulate; default 10.}
}
\value{
Dataframe of number of snailse for each allele for each pop for each gen.
}
\description{
Generate multiple generations of simulations of genetic drive
}
\examples{
sim_snails()
sim_snails(100, 10, 100)

}
